Embed recoverable and unrecoverable signature
